feat(bcgov/storybook): update navigation
change to mdx formate

#294
